Remove dead code and refactor DI
